feat: split single routes with satellites
