+++ src/frontend/src/routes/(single)/(satellites)/analytics/+page.svelte
