+++ src/frontend/src/routes/(single)/(standalone)/mission-control/+layout.svelte
