+++ src/frontend/src/routes/(single)/(standalone)/mission-control/+page.svelte
