+++ src/frontend/src/routes/(single)/(standalone)/monitoring/+layout.svelte
