+++ src/frontend/src/routes/(single)/(standalone)/monitoring/+page.svelte
