+++ src/frontend/src/routes/(single)/(standalone)/preferences/+layout.svelte
