+++ src/frontend/src/routes/(single)/(standalone)/preferences/+page.svelte
