+++ src/frontend/src/routes/(single)/(standalone)/wallet/+layout.svelte
