+++ src/frontend/src/routes/(single)/(standalone)/wallet/+page.svelte
